+++ lib/index.ts
@@ -7,3 +7,5 @@ import JSONDriver from 'drivers/Json';
 
 export { Database, DatabaseManager, Driver, BSONDriver, JSONDriver };
 export default Database;
+
+// TODO - Add DatabaseServer.